Move value extraction of `size` prop to the component level
